Normalize dashes in wheel package names (fixes #27)
This patch normalizes package names in wheels such that `pip install
hello-world` will install a wheel if available. Given a package with a
dash in the name, like `hello-world`, `python setup.py bdist_wheel` will
create a wheel with a package name of `hello_world`. Unfortunately,
running `pip install hello-world` will not look for a `hello_world`,
only `hello-world`, and will never find a wheel with that package.