Ignore .repo directories created by the git-repo tool
git-repo [1] is used to manage multiple Git repositories. The .repo
directory contains the consolidated Git histories of the managed projects
(which are symlinked from the individual projects' .git repositories), so
we should ignore the .repo directory just like the .git directory.

However, the .repo directory also contains the source code to the git-repo
tool itself. One might argue that this source code should be picked up
during a scan, but then again the code of git-repo is not integrated into
the deliverable / what is being built from the source code as git-repo is
just a VCS tool, so a user is probably not interested in any license
findings in git-repo.

[1] https://code.google.com/p/git-repo/