The GPL 2.0.4
This is the last version of the GPL text as published by the FSF.
Let's call it the GPL 2.0.4
This variation was published around about the time of the FSF released
the GPL 3 in July 2007.
http://web.archive.org/web/20070716031727/http://www.gnu.org/licenses/old-licenses/gpl-2.0.txt
It is found live here
https://www.gnu.org/licenses/old-licenses/gpl-2.0.txt
https://www.gnu.org/licenses/old-licenses/gpl-2.0.html
It refers to the Franklin Street address and to the GNU Lesser General
Public License everywhere both in the text and HTML formats.

Signed-off-by: Philippe Ombredanne <pombredanne@nexb.com>