Update the GPL license with notes about which text version it is.
Remove the duplicate gpl-2.0_104.RULE

This is the last version of the GPL text as published by the FSF.
This variation was published around about the time of the FSF released
the GPL 3 in July 2007.
    http://web.archive.org/web/20070716031727/http://www.gnu.org/licenses/old-licenses/gpl-2.0.txt
It is found live here
    https://www.gnu.org/licenses/old-licenses/gpl-2.0.txt
    https://www.gnu.org/licenses/old-licenses/gpl-2.0.html
It refers to the Franklin Street address and to the GNU Lesser General
Public License everywhere both in the text and HTML formats.
There are four other variations that were published over the years. You
can find these in the GPL rules as gpl-2.0_100.RULE, gpl-2.0_101.RULE,
gpl-2.0_102.RULE and gpl-2.0_103.RULE. Their corresponding YAML data
explains their history and provides archive.org pointer to their texts.

Signed-off-by: Philippe Ombredanne <pombredanne@nexb.com>