When configuring, print which config is being used even in quiet mode
As there is a subtle difference when running ./configure directly (dev
config is being used) and ./configure being run by the scancode script
(dev config not being used), print the config being usd in any case.
This is usuall is just two lines and does not really clutter the output.