Docs: Added workaround for testing on RedHat OSs
Since Scancode binaries are built on Debian based machines, and the
prebuilt dependencies are named differently on RedHat based machines
py.test will not be able to find the libbz2.so library on those
machines.
See https://github.com/nexB/scancode-toolkit/issues/443

Added a note about how to work around this issue by symbolically
linking the existing libbz2.so on the filesystem to the expected
name.

This should be removed once
https://github.com/nexB/scancode-toolkit/issues/469 is solved

Signed-off-by: Nisha K <nishak@vmware.com>