Return minimal type information #802
AboutCode Manager and other tools need to know if a path is a file or
directory at the minimum. This information was previously returned
only with the `--info` option. We now always return minimal type
file information with a scanned file for proper downstream processing
of scan results by tools like the AboutCode Manager.

Signed-off-by: Yash Nisar <yash.nisar@somaiya.edu>