Rename data.json to data.js
The file data.json in the html-app isn't a valid json. It contains a js
variable containing a valid json, so basically the file ending should
be js.

When I want to open the html app via the Jenkins HTML publisher plugin,
I get the following error in the Chrome console:
Refused to execute script from 'https://.../report_files/data.json'
because its MIME type ('application/json') is not executable, and strict
MIME type checking is enabled.

This commit fixes that issue.

Signed-off-by: mueller-ma <mueller-ma@users.noreply.github.com>