Add new plugin to improve classification
Some packages have extra root-like file locations (such as META-INF
directory in a JAR). This allow to account for these and use that
when classifying files. This is a post-scan plugin that comes after
a possible package scan and should also be before any of the
summarization plugins that would need the classification.

Signed-off-by: Philippe Ombredanne <pombredanne@nexb.com>