Treat digit lines specially in license detection #1404
Some code or data files may contain long runs of numbers as data.
These are no super interesting for license detection so we treat them
specially as if they were empty when breaking a query in smaller
query runs. This improves the speed at which such files are scanned
significantly.
Also we now skip doing a sequence match for a query run if there is
no high matchable tokens left. Since it would not be matchable as a
sequence anyway there is no reason to try to match it further after
automaton matching.

The combination of these tow optimisations means that Vorbis scanning
time is more than halved (from ~30 secs down to ~13 secs)

Signed-off-by: Philippe Ombredanne <pombredanne@nexb.com>