Validate that test method names are unique
We could have cases of two files names that -when reduced to a python
safe name for use as a test method name- would no longer be unique and
when a test method was attached to a test class, only one would stick.

This meant that some tests were not running at all or that some tests
were failing randomly depending on which machine they ran on.

This patch ensures that:
1. we generate names that are more likely unique
2. we check all generated names are unique and fail if not

Signed-off-by: Philippe Ombredanne <pombredanne@nexb.com>