Remove redundant hash match license tests
Several tests are not useful since they check an exact match and are
bringing nothing new. We just kept a few  and removed the other ones.
A hash match test is kinda tautological and does not excercise anything
in the license matching algorithms. Even if runs fast it still runs for
nothing.

Signed-off-by: Philippe Ombredanne <pombredanne@nexb.com>