Used fixed list of legalese words
We now use the "legalese" data set throughout, switching the tests
on >= len_junk by a test on < len_legalese to determine if a token is
legalese or junk. This fixed set of legalese words is better suited and
easier to maintain than the unbounded set of "junk" words that was used
previously.

This change has impacts throughout the license detection code.
Some tests were removed as no longer relevant (such as testing the
posting of high tokens).

Signed-off-by: Philippe Ombredanne <pombredanne@nexb.com>