Update configure scripts for Python 2 and 3 #295
On Windows, configure.bat now accepts a --python <path to exe> arg to
allow setting up the project with a Python 3 exe path.

The configure.py script now includes activate_this.py which is not
part of the Python 3 "venv" module. It has also been updated
significantly to support working from paths with spaces (which is the
norm on Windows for Python 3 installation paths).

Signed-off-by: Philippe Ombredanne <pombredanne@nexb.com>