Correct license text collection bug
The license text was truncted in some cases because of unicode
diacritics behaving slightly differently in regular expression when
lowercased or or not.

This led to certain lowercased diacritics not being the same with the
regular query tokenizer an the matched text tokenizer and leading to
an off by one error and a truncation of the collected license text.

This fix changes how the check is done to verify if a token is a known
token or not.

This also introduce a "highlight" argument to the
LicenseMatch.matched_text() methods and related functions


Signed-off-by: Philippe Ombredanne <pombredanne@nexb.com>