Handle Turkish diacritics in license text #1872
Somehow the some characters have weird behaviours and decompose to more
than one character when going from uppercase to lowercase. This applies
when following strictly Unicode (which is now the acse in Python 3) and
quite specifically in Turkish and Azeri. The code was design to crash
if we ever met such case and it did recently as part of #1872

This fixes the issue encountered when reconstructing the matched text
of license matches by returning lowercased versions of the words that
contain such a rare diacritics.

Signed-off-by: Philippe Ombredanne <pombredanne@nexb.com>