Include prefix in SPDX license matches #2007
We now include the SPDX-license-Identifier prefix (and its English
spelling) in the matched text of an "SPDX-license-Identifier" tag.
Therefore we avoid the cases where a plain text rule would partially
overlap with an SPDX rule and both would be returned (when only one
should be returned).

Along the way comments and code clarity was improved with small
refactorings and additions.
Removed some extraneous tracing code.

Signed-off-by: Philippe Ombredanne <pombredanne@nexb.com>