Restore progress reporting to extractcode #1996
Do not consume extract events generators in a list. This prohinited
the progressive dispatch of "events" being generated to provide
progress display in the CLI. Instead we now yield events again as they
are created and accumulate events for leter reprocessing if the
option to replace originals has been selected.

Signed-off-by: Philippe Ombredanne <pombredanne@nexb.com>