licensecheck: add LRE-based scan, with only MIT license
This adds a new function Scan that is meant to replace Cover.
Scan has a different API and different semantics but is much faster
and more scalable because it uses a DFA and insists on more
precise matches, avoiding the cost (and imprecision) of Cover's fuzzy matching.

Once Scan is capable of taking over completely for Cover,
we will remove Cover.