Improve license rules validation
Break the Rule object in BasicRule that can be constructed without files
and Rule proper that requires a text and data file. This ensures that
rules created without files can also be validated.

Ensure that we can catch more model inconsistencies and move validation
proper to a BasicRule.validate() method.

Signed-off-by: Philippe Ombredanne <pombredanne@nexb.com>