Fix logic error in is_empty
The logic in is_empty is reversed - the function tests for *emptiness*, so a falsey "clues" object should actually return True (it *is* empty). 

This prevents an error that manifests at line 241, where attributes of the ignorables object are used, but the object is None. The reversed logic test actually fails at 238, where the function should return (there are no ignorables), but the default None prevents the "and" operator from returning True, so the function continues at 242.