Use common function to get HTTP content and headers
HTTP HEAD request do not work well with S3 which is used by GitHub:
they do not return the exact headers that would be returned with a GET
in particula it is missing the ETag. This uses a "stream" GET request to
get headers at all times and get content when needed.

Also do minor refactorings.

Signed-off-by: Philippe Ombredanne <pombredanne@nexb.com>