restore git diff functionality
all aboutcode repositories have "* binary" in .gitattribute

This has the side effet of disabling the git diff functionality from the commandline.

git diff will always refuse to report diff and say the binary files differ

problem is tests are requiring some of the dataset is kept as is, so we mark
all the **/data/** as binary, as well as certain other files.

Signed-off-by: Pierre Tardy <tardyp@gmail.com>