Test Debian copyright for full match details
The Debian copyright tests expected data were just a list of items
without an attribute name telling what each item was.

This fixes this by making the expected YAML results a dictionary rather
than a list. It also adds the full match details to the expected results such that we can better assess what failed and how and can better
update the tests as we improve license detection.

Signed-off-by: Philippe Ombredanne <pombredanne@nexb.com>