Consider license rule length in set rankings #2635
Prior to this fix, in the set matching step of license detection,
the ranking (and deduplication) of candidates eligible for actual full
sequence matching was ignoring the rule length and only privileging
the matched length, ressemblance and containment. In case of a tie
between two candidate rules it makes sense to privilege the longest
rule which is what this commit does.

Reported-by: Peter Gardfjäll <peter.gardfjall.work@gmail.com>
Signed-off-by: Philippe Ombredanne <pombredanne@nexb.com>