Use tag file to restart release tests
We now create a tag file after each succesful test of a release archive
and this mean we can restart from the last successful tests and not
restest everything. Azure and/or the network are rather flaky lately
so this way we can run the full release matrix even if it means
restarting the test command several times.

Signed-off-by: Philippe Ombredanne <pombredanne@nexb.com>