Add key phrase filtering to licensedcode
Test `test_LicenseMatch_matches_only_when_key_phrases_are_present` demonstrated the main purpose of this change.

Without setting `key_phrase_spans` on `r1` it would match the `querys` query string.
This would be incorrect because to a human `querys` is obviously an Apache License, and the rule an MIT license.

When `key_phrase_spans` is set, in this example to `Span(4)` which matches `MIT` an exact match is required for that span.
Since for the query `Span(4)` matches `Apache` the `key_phrase_spans` can not be found within the `ispan` of the match, and thus be filtered out.