Add key phrase parsing to each rule
It the moment I've implemented the code inside the `Rule` itself, but only initialize and set in during indexing since I am not sure what the best location for this would be performance wise.

I've used the tokens returned by `tokens()` in order to get the correct token positions, because some words (STOPWORDS) are filtered out and AFAIK are not considered during matching.