Add lexer and style for LilyPond (#1845)
* Add lexer and style for LilyPond

GNU LilyPond is a text-based music typesetter. Because its concepts
are completely different from programming languages, this adds
a special-purpose to highlight the special tokens.

The SchemeLexer is expanded to gain the ability of detecting when
one expression stops. LilyPondLexer subclasses SchemeLexer.

Builtins (the most important part) are generated with a script
put in external/ (as was already done for Lasso).

As part of this change, the CPSALexer is made no longer to
inherit from SchemeLexer. The inheritance was unused anyway.

* Fixup: translators names have underscores

* Fixup: avoid duplicate builtins

* Fixup: update goldens

* Fixup: typo

* Fixup: add missing tokens to style

* Fixup: update lexer comments

* Fixup: Tentative style adjustments

* Fixup: move test file to examplefiles/

* Fixup: miscellaneous fixes (to be finished)

* Fixup: escape all braces

* Fixup: use Text, not Whitespace

* Fixup: fixes for lexing and style

* Fixup: update goldens!

* Fixup: also test alist assignments

* Fixup: recognize escape sequences in strings

* Fixup: use Comment.Single

* Fixup: Whitespace, not Text!

* Fixup: fix pitch parsing

* Fixup: update comment

* Fixup: remove redundant re.UNICODE