Improve handling of continuous and key phrases
Remove non-continuous filter and use instead the key phrases filter,
as a rule that must be matched continuously is the same as a rule with
a single key phrase covering all its text

Track stopwords on the index side in each rule, and use these to ensure
that when keyphrases are present and contain stopwords we check that the
stopwords are the same on the query and index side of a match to
consider the keyphrases matches.

For stopwords introduced a new index tokenizer
"index_tokenizer_with_stopwords".

Sort stopwords for easier search in the future.

Streamline keyphrase tokenizer and run it in Rule.tokens()

Remove query.Query.stopwords_span that was not used
Remove match.LicenseMatch.qcontains_stopwords() that was not used
Remove match.filter_non_continuous_matches() no longer used

Add new and improved rules, and requalify some non-english short rules
as bona-fide rules.

Apply multiple minor refactoring, doc, and formatting refinements

Signed-off-by: Philippe Ombredanne <pombredanne@nexb.com>