Recognize either app or system packages or both
This way it is possible to focus a scan only on some packages types
and not always boil the ocean.

For this we now have two lists of DatafileHandlers: one for system
packages and one for application packages.

The new UX is to have a change in the command line options:
--package detects only application packages (which basically is what it
was doing in the previous release)
--system-package is new and detects only system packages.
These two options can be combined to detect all package types.

Several functions now accept new "application=True, system=False"
arguments


Signed-off-by: Philippe Ombredanne <pombredanne@nexb.com>