Improve release scripts #3040
Some release archves where missing critical wheels.
The solution is to:

* ensure that we have effectively missing wheels published and made
available from our PyPI repo when building a release

* Improve the fetch_thirdparty script to fall back on fetching a
source distribution when no wheel is found. Fail loudly if neither a
wheel not an sdist is found. Also accept special cases of packages with
no wheels and packages with no sources.

* Ensure that some source-only packages are always included as sources

* Add release smoke tests and run then for each wheel, sdist and app
  archive on supported OS and Python versions combinations

* Move Linux-only requirements to theur own requirement file.

* Enable Linux-only package inspectors on Linux only.

* Run base tests on macos 12

Signed-off-by: Philippe Ombredanne <pombredanne@nexb.com>