Add unknown license reference to package dereferencing #2965 #1379
There are unknown license statements like "This file is the same
license as the package django" which refers to a package which this
file is a part of. This is fixed by extending the dereferencing logic
to look for packages that the file belongs to and using the detected
licenses from the package.

Signed-off-by: Ayan Sinha Mahapatra <ayansmahapatra@gmail.com>