Update test file to non-proprietary license variant
Switch to use a version that's under
https://scancode-licensedb.aboutcode.org/sun-bsd-no-nuclear.html

The long history is this:

This is code with a complex history as it was also published under a proprietary license at:
https://web.archive.org/web/20000823095253/http://java.sun.com/products/jfc/tsc/articles/treetable2/src/MergeSort.java
and at:
https://web.archive.org/web/20000823094850/http://java.sun.com/products/jfc/tsc/articles/treetable1/src/MergeSort.java
and is part of code that was relicensed by Sun following some issues raised many years ago.

The earliest relicense is at
https://web.archive.org/web/20030709130712/http://java.sun.com:80/products/jfc/tsc/articles/treetable1/src/MergeSort.java

The detailed history is found foremost in this Mozilla bugzilla:
https://bugzilla.mozilla.org/show_bug.cgi?id=419491and
And also in Mozilla Rhino history at:
https://github.com/mozilla/rhino/blob/2c8ae233cda5d6129a140e5572a6dda148afc039/toolsrc/org/mozilla/javascript/tools/debugger/build.xml#L40
and:
https://github.com/mozilla/rhino/commit/2e746cd5fb4de9a641b10550909997acb508d324

The gist of this is that all versions prior to Mozilla Rhino prior to
1.7R1 contained Sun proprietary binary code injected at build time that
was fetched from code listed above and attached to the article at:
https://web.archive.org/web/20020520083650/http://java.sun.com/products/jfc/tsc/articles/treetable2/

This meant that use of those versions would have required permission
from Sun (now Oracle) to use and distribute these Sun proprietary
components. Sun subsequently released and updated the code under a
BSD style license, and Rhino 1.7R1 and later further vendored the code
now at:
https://github.com/mozilla/rhino/tree/6387bac4281ba1c4a8050cd055071615593d214c/toolsrc/org/mozilla/javascript/tools/debugger/treetable

Ironically I am tied to the history of this relicensing somehow as seen
in comments in at
https://bugzilla.mozilla.org/show_bug.cgi?id=419491

More details for the posterity: Circa 1998, Sun employees published an
article with some source code attached. The code was under a Sun
proprietary license. Circa 2008, at the request of the Mozilla Rhino
project (which was using that code) Sun relicensed part of that article
attached code under a new BSD license (as explained in
https://bugzilla.mozilla.org/show_bug.cgi?id=419491 ).
This code was also already relicensed under an earlier version of the
article in 2003 under a BSD-style license.

This test file (used for testing line endings on old MacOS) was using
originally the proprietary-licensed variant.
This is NOT part of the released commoncode wheels deployed at runetime
and is not an issue per-se since this code was also licensed by SUN
under this BSD-style license
https://scancode-licensedb.aboutcode.org/sun-bsd-no-nuclear.html

This file is therefore updated here to use the
sun-bsd-no-nuclear-licensed version and will be deleted afterwards as
this is no longer used in commoncode as the read_text_file() function
it tested as been removed.

Reference: https://web.archive.org/web/20030709130712/http://java.sun.com:80/products/jfc/tsc/articles/treetable1/src/MergeSort.java
Signed-off-by: Philippe Ombredanne <pombredanne@nexb.com>