Add maven license detection updates
Add a new config variable for datafile handlers, which if enabled
will run a package-ecosystem specific implementation of license
detection. Added a function to implement this for maven which only keeps
the relevant information and runs license detection on the whole
license statement and not on the respective values/attributes
one-by-one. Enabled this for maven to improve license detection there.

Signed-off-by: Ayan Sinha Mahapatra <ayansmahapatra@gmail.com>