Add new attribute `matched_text_diagnostics`
This commit adds a new attribute for license text diagnostics
which will be added when the CLI option `--license-text-diagnostics`
is used, in contrast to earlier behaviour, where this diagnostics
matched text used to overwrite the text in `matched_text`.

Also makes sure top-level license/package summarizations have
matched text and diagnostics correcttly only when the respective
CLI options are used.

Signed-off-by: Ayan Sinha Mahapatra <ayansmahapatra@gmail.com>