Detect distro files in system package scans
In docker image/other rootfs scans with SCIO we are always
scanning for system packages before we scan for application
packages, and we were not being able to use the distro
information as this was being detected in application
package scans happening later. So in this commit we are
also adding os-release file parser to system package handler.

Signed-off-by: Ayan Sinha Mahapatra <ayansmahapatra@gmail.com>