Compute checksums from binary stream
Scanning a large file may exhaust memory and file because
commoncode.hash loads the whole file in memory. With this commit, the
files are loaded in chunks so we do not run out of memory on very large
files, such as VM images.

Reference: https://github.com/aboutcode-org/commoncode/issues/72
Signed-off-by: Philippe Ombredanne <pombredanne@nexb.com>