feat(OutputHTML): Implement fallback for license references table
The `--license-references` cli argument enables collection and storge
of license reference data in output files. So as long as it was present
as a Post-Scan option the `license_references` should not be empty.
In such cases, the license reference table shall be filled with the
already stored data.

If however the `license_references` list is empty, the licenses should
be collected from the `license_db` as a fall back, similar to the legacy
license collection pre 49e7d89.

Fixes #4101.

Signed-off-by: Jens Keim <jens.keim@forvia.com>