Improve Maven extracted_license storage and detection
We were only storing the list of attributes but the now "license" top
level elements. Because of this a maven tag with only name: mit would
not be detected properly as there is no rule for this and this would
yield too many false positive. Now we have a structure that is a list
of license as {license: {name: mit, url: foo: comments: bar}}

This ensures a more accurate detection. With this also regenerate all
Maven related tests, and also ensure we filter early empty and unused
license attributes (such as the "distribution")

Also add new tests and apply mild refactoring for maven license
detection for clarity.

Remove redundant tests and simplify some other tests for clarity.

Signed-off-by: Philippe Ombredanne <pombredanne@nexb.com>